# expense_tracker_backend/interfaces/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Expense Tracker API",
    "version": "1.0.0",
    "description": "A comprehensive expense tracking API with magic link authentication, expense management, and user profile features.",
    "contact": {
      "name": "API Support",
      "email": "support@expense-tracker.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
      }
    },
...
  },
...
  "paths": {
    "/api/auth/magic-link": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Request magic link for email authentication",
        "description": "Send a magic link to the provided email address for passwordless authentication",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address",
                    "example": "user@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Magic link sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Magic link sent to your email"
                    },
                    "magicLink": {
                      "type": "string",
                      "description": "Only included in development mode",
                      "example": "http://localhost:3001/auth/verify?token=abc123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}